Version 0.3.3 - Setup page complete, user can now set profile picture, description and location from setup page. New PHP file for account information update. Any account information can be updated using single script.

+++ Resources/Php/Db/database.sql
@@ -17,10 +17,10 @@ Create Table BikeMaster.Users(
     	Not Null,
 	ProfilePictureUrl varchar(255),
 	Description varchar (511),
-	 /*-90-90 deg. Precision 4 decimal places (11.1m)*/
-    Latitude double(6, 4),
-    /*-180-180 deg. Precision 4 decimal places (11.1m)*/
-    Longitude double(7, 4),
+	 /*-90-90 deg. Precision 3 decimal places (111.1m)*/
+    Latitude double(6, 3),
+    /*-180-180 deg. Precision 3 decimal places (111.1m)*/
+    Longitude double(7, 3),
     RegistrationTime int
 		Unsigned
 		Default Unix_Timestamp()
